Fix ecash error when redeeming a token
Fixes the "blinded message already signed" error which would happen when
CDK was trying to get an already expired secret re-signed by the mint.

By "restoring" the proofs we get back in sync with the mint and we know
which secrets have been spent. We assume the same key is not being used
anywhere else, so we only run the restore when we don't have any proofs
stored locally for a given (mint_url, unit) pair.