Fix subscription regression
When the subscription were moved out of the pending request context a
few regressions were introduced. The biggest one was #93, but there were
a few other issues related to balance checks, calendar management and
missing activities/payment status.

This attemps to fix the regressions, but there's still work to do to
make this work well.